requirements.txt
google-generativeai
PyPDF2
langchain
reportlab
python-dotenv
langchain-text-splitters
